auth_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
sqlalchemy
python-dotenv
python-multipart
